+++ packages/site/content/components/do-and-dont.mdx
@@ -14,11 +14,11 @@ import {DoDontContainer, Do, Dont} from '@primer/doctocat-nextjs'
 ```jsx
 <DoDontContainer>
   <Do>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you should do</Caption>
   </Do>
   <Dont>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you shouldn't do</Caption>
   </Dont>
 </DoDontContainer>
@@ -28,11 +28,11 @@ Will render:
 
 <DoDontContainer>
   <Do>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you should do</Caption>
   </Do>
   <Dont>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you shouldn&apos;t do</Caption>
   </Dont>
 </DoDontContainer>
@@ -44,11 +44,11 @@ Use the `stacked` option to vertically stack the Do and Don&apos;t images.
 ```jsx
 <DoDontContainer stacked>
   <Do>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you should do</Caption>
   </Do>
   <Dont>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you shouldn&apos;t do</Caption>
   </Dont>
 </DoDontContainer>
@@ -56,11 +56,11 @@ Use the `stacked` option to vertically stack the Do and Don&apos;t images.
 
 <DoDontContainer stacked>
   <Do>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you should do</Caption>
   </Do>
   <Dont>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you shouldn&apos;t do</Caption>
   </Dont>
 </DoDontContainer>
