+++ packages/site/content/components/video.mdx
@@ -2,3 +2,19 @@
 title: Video
 description: Adding video to your documentation pages
 ---
+
+Use the `<video>` element for presenting video content.
+
+<video width="600" controls>
+  <source src="/video/example.mp4" type="video/mp4" />
+  <track kind="captions" src="/video/example.vtt" default />
+  Your browser does not support the video tag.
+</video>
+
+```HTML
+<video width="100%" controls>
+  <source src="/video/example.mp4" type="video/mp4" />
+  <track src="/video/example.vtt" default />
+  Your browser does not support the video tag.
+</video>
+```
