+++ packages/site/content/content-examples/kitchen-sink.mdx
@@ -1,13 +1,13 @@
 ---
 title: Kitchen sink
 description: This is a quick introduction to what is going to be developed here. That can be a chunky paragraph that expands quite a bit of context.
-image: https://via.placeholder.com/1000x450/f1f1f1/f1f1f1.png
+image: /images/placeholder-1200x630.png
 image-alt: Image description
 action-1-text: Primary action
 action-1-link: /
 action-2-text: Secondary action
 action-2-link: /content-examples/kitchen-sink
-keywords: ['accessibility', 'introduction', 'simple']
+keywords: ['example']
 ---
 
 ## Heading 2
@@ -45,7 +45,7 @@ codebase and identify vulnerabilities before they reach production.
 
 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
 
-![Placeholder, blank image](https://via.placeholder.com/1000x350/d3d9df/d3d9df.png)
+![Placeholder, blank image](/images/placeholder-1200x630.png)
 
 ---
 
@@ -53,11 +53,11 @@ Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor i
 
 <DoDontContainer stacked>
   <Do>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you should do</Caption>
   </Do>
   <Dont>
-    <img src="https://via.placeholder.com/1000x350/d3d9df/d3d9df.png" alt="" />
+    <img src="/images/placeholder-1200x630.png" alt="" />
     <Caption>Example of what you shouldn't do</Caption>
   </Dont>
 </DoDontContainer>
