+++ packages/site/content/content-examples/simple.mdx
@@ -1,7 +1,12 @@
 ---
 title: Simple
 description: A simple page with plain markdown
-keywords: ['kitchen sink', 'introduction', 'simple']
+keywords: ['example']
+related:
+  - title: 'Typography in Product UI'
+    href: 'https://primer.style/foundations/typography'
+  - title: 'Monaspace'
+    href: 'https://monaspace.githubnext.com/'
 ---
 
 ## Arva qua ferarum victa
