+++ packages/site/content/getting-started/introduction.mdx
@@ -4,22 +4,20 @@ description: Learn how to create a new site with Doctocat
 keywords: ['accessibility', 'introduction', 'simple']
 ---
 
-<Button>Test</Button>
-
 This guide will walk you through creating, customizing, and deploying a new documentation site powered by Doctocat on [Next.js](https://nextjs.org/).
 
 ## 1. Create a new Next.js installation
 
-Start by creating a new Next.js project using the [Next.js CLI](https://nextjs.org/docs/getting-started/installation):
+Start by creating a new Next.js project using the [Next.js CLI](https://nextjs.org/docs/getting-started/installation).
 
-Note: Select "No" for question: "Would you like to use App Router?"
+Select "Yes" for the question: "Would you like to use App Router?"
 
 ## 2. Install Doctocat (Next.js) and relevant dependencies
 
 Next, install Doctocat and its dependencies:
 
 ```bash
-npm install --save @primer/doctocat-nextjs next@v14.0 eslint-config-next@v14.0
+npm install --save @primer/doctocat-nextjs next
 ```
 
 ## 3. Update your Next.js configuration
@@ -31,14 +29,11 @@ Update your next.config.js file to use the Doctocat theme:
 /** @type {import('next').NextConfig} */
 import withDoctocat from '@primer/doctocat-nextjs/doctocat.config.js'
 
-const nextConfig = withDoctocat({
-  reactStrictMode: true,
-  publicRuntimeConfig: {
-    siteTitle: 'Doctocat site', // add your own site title here as an environment variable
-  },
-})
-
-export default nextConfig
+export default {
+  ...withDoctocat({
+    transpilePackages: ['@primer/doctocat-nextjs'],
+  }),
+}
 ```
 
 ### Optional: GitHub Pages configuration
@@ -51,7 +46,15 @@ If you plan to deploy your site to GitHub Pages, you'll need to add the followin
 }
 ```
 
-## 4. Add the Doctocat stylesheets
+## 4. Add required environment variables
+
+```shell
+NEXT_PUBLIC_SITE_TITLE = Documentation # required: for site metadata and header title
+NEXT_PUBLIC_REPO = https://github.com/org/repo # required: for edit links
+NEXT_PUBLIC_REPO_SRC_PATH = # optional:  path for monorepos that do not store .mdx in the project root
+```
+
+## 5. Add the Doctocat stylesheets
 
 Add the required Doctocat stylesheets to your `pages/_app.(tsx|jsx)` file:
 
@@ -66,7 +69,7 @@ function App({ Component, pageProps }: AppProps) {
 export default App;
 ```
 
-## 5. Update homepage
+## 6. Update homepage
 
 Rename `pages/index.tsx` to `pages/index.md` and replace the file contents with:
 
@@ -91,7 +94,7 @@ Important:
 - You must use `.md` or `.mdx` extensions for your documents.
 - Folders must contain an index.md or index.mdx file for breadcrumbs to work correctly. The contents of this file can be left empty
 
-## 6. Update Frontmatter on all pages
+## 7. Update Frontmatter on all pages
 
 Side navigation for your site is generated automatically from the frontmatter of your pages. To add a page to the side navigation, you must add a `title` and `description` to the frontmatter of your page. For example:
 
@@ -102,7 +105,7 @@ description: My page description
 ---
 ```
 
-## 7. Deploy to GitHub Pages
+## 8. Deploy to GitHub Pages
 
 After you've customized the content of your site, you're ready to deploy. There are many ways to deploy your site, but we currently use [GitHub Pages](https://pages.github.com/) for most of our sites, and have found it to be an easy way to get sites up and running quickly.
