+++ packages/site/content/index.mdx
@@ -2,6 +2,7 @@
 title: Homepage
 description: That
 keywords: test, test2
+layout: custom
 ---
 
 import HomepageComponent from './index.tsx'
