+++ packages/theme/components/content/dos-and-donts/DosAndDonts.module.css
@@ -50,7 +50,7 @@
 
 .content img {
   max-width: 100%;
-  margin-block-end: 0;
+  margin-block-end: 0 !important;
 }
 
 .indentedContent {
