+++ packages/theme/components/layout/related-content-links/RelatedContentLinks.tsx
@@ -17,7 +17,9 @@ export function RelatedContentLinks({links}: RelatedContentLinksProps) {
   if (!links?.length) return null
   return (
     <div className="custom-component">
-      <Heading as="h2">Related content</Heading>
+      <Heading as="h2" size="subhead-large">
+        Related content
+      </Heading>
       <UnorderedList className={styles.list}>
         {links.map(page => (
           <UnorderedList.Item key={page.route}>
